Adding unit test
