+++ tests/unit/datarobot_drum/drum/test_drum.py
@@ -7,153 +7,31 @@
 #
 
 import os
-import sys
 from copy import deepcopy
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from typing import List
 from unittest.mock import ANY, PropertyMock, patch
 
 import pandas as pd
 import pytest
-import yaml
 from datarobot_drum.drum.adapters.cli.drum_fit_adapter import DrumFitAdapter
 from datarobot_drum.drum.adapters.model_adapters.python_model_adapter import PythonModelAdapter
-from datarobot_drum.drum.args_parser import CMRunnerArgsRegistry
 from datarobot_drum.drum.drum import (
     CMRunner,
     create_custom_inference_model_folder,
     output_in_code_dir,
 )
-from datarobot_drum.drum.enum import MODEL_CONFIG_FILENAME, RunLanguage
+from datarobot_drum.drum.enum import RunLanguage
 from datarobot_drum.drum.language_predictors.python_predictor.python_predictor import (
     PythonPredictor,
 )
-from datarobot_drum.drum.runtime import DrumRuntime
 from datarobot_drum.drum.utils.structured_input_read_utils import StructuredInputReadUtils
 from mlpiper.pipeline.executor import Executor
-
-
-def set_sys_argv(cmd_line_args):
-    # This is required because the sys.argv is manipulated by the 'CMRunnerArgsRegistry'
-    cmd_line_args = cmd_line_args.copy()
-    cmd_line_args.insert(0, sys.argv[0])
-    sys.argv = cmd_line_args
-
-
-def get_args_parser_options(cli_command: List[str]):
-    set_sys_argv(cli_command)
-    arg_parser = CMRunnerArgsRegistry.get_arg_parser()
-    CMRunnerArgsRegistry.extend_sys_argv_with_env_vars()
-    options = arg_parser.parse_args()
-    CMRunnerArgsRegistry.verify_options(options)
-    return options
 
 
 @pytest.fixture
 def module_under_test():
     return "datarobot_drum.drum.drum"
-
-
-@pytest.fixture
-def this_dir():
-    return str(Path(__file__).absolute().parent)
-
-
-@pytest.fixture
-def target():
-    return "some-target"
-
-
-@pytest.fixture
-def model_metadata_file_factory():
-    with TemporaryDirectory(suffix="code-dir") as temp_dirname:
-
-        def _inner(input_dict):
-            file_path = Path(temp_dirname) / MODEL_CONFIG_FILENAME
-            with file_path.open("w") as fp:
-                yaml.dump(input_dict, fp)
-            return temp_dirname
-
-        yield _inner
-
-
-@pytest.fixture
-def temp_metadata(environment_id, model_metadata_file_factory):
-    metadata = {
-        "name": "joe",
-        "type": "training",
-        "targetType": "regression",
-        "environmentID": environment_id,
-        "validation": {"input": "hello"},
-    }
-    yield model_metadata_file_factory(metadata)
-
-
-@pytest.fixture
-def output_dir():
-    with TemporaryDirectory(suffix="output-dir") as dir_name:
-        yield dir_name
-
-
-@pytest.fixture
-def fit_args(temp_metadata, target, output_dir):
-    return [
-        "fit",
-        "--code-dir",
-        temp_metadata,
-        "--input",
-        __file__,
-        "--target",
-        target,
-        "--target-type",
-        "regression",
-        "--output",
-        output_dir,
-    ]
-
-
-@pytest.fixture
-def score_args(this_dir):
-    return [
-        "score",
-        "--code-dir",
-        this_dir,
-        "--input",
-        __file__,
-        "--target-type",
-        "regression",
-        "--language",
-        "python",
-    ]
-
-
-@pytest.fixture
-def server_args(this_dir):
-    return [
-        "server",
-        "--code-dir",
-        this_dir,
-        "--address",
-        "allthedice.com:1234",
-        "--target-type",
-        "regression",
-        "--language",
-        "python",
-    ]
-
-
-@pytest.fixture
-def runtime_factory():
-    with DrumRuntime() as cm_runner_runtime:
-
-        def inner(cli_args):
-            options = get_args_parser_options(cli_args)
-            cm_runner_runtime.options = options
-            cm_runner = CMRunner(cm_runner_runtime)
-            return cm_runner
-
-        yield inner
 
 
 @pytest.fixture
